--- a/go.mod
+++ b/go.mod
@@ -109,6 +109,7 @@ require (
 	go.uber.org/goleak v1.2.1
 	golang.org/x/exp v0.0.0-20230817173708-d852ddb80c63
 	golang.org/x/sync v0.3.0
+	gonum.org/v1/gonum v0.14.0
 	modernc.org/sqlite v1.20.3
 )