--- a/go/pools/resource_pool.go
+++ b/go/pools/resource_pool.go
@@ -35,41 +35,12 @@ import (
 )
 
 type (
-	IResourcePool interface {
-		Close()
-		Name() string
-		Get(ctx context.Context, setting *Setting) (resource Resource, err error)
-		Put(resource Resource)
-		SetCapacity(capacity int) error
-		SetIdleTimeout(idleTimeout time.Duration)
-		StatsJSON() string
-		Capacity() int64
-		Available() int64
-		Active() int64
-		InUse() int64
-		MaxCap() int64
-		WaitCount() int64
-		WaitTime() time.Duration
-		IdleTimeout() time.Duration
-		IdleClosed() int64
-		MaxLifetimeClosed() int64
-		Exhausted() int64
-		GetCount() int64
-		GetSettingCount() int64
-		DiffSettingCount() int64
-		ResetSettingCount() int64
-	}
-
 	// Resource defines the interface that every resource must provide.
 	// Thread synchronization between Close() and IsClosed()
 	// is the responsibility of the caller.
 	Resource interface {
-		Close()
 		Expired(time.Duration) bool
-		ApplySetting(ctx context.Context, setting *Setting) error
-		IsSettingApplied() bool
-		IsSameSetting(setting string) bool
-		ResetSetting(ctx context.Context) error
+		Close()
 	}
 
 	// Factory is a function that can be used to create a resource.
@@ -78,12 +49,6 @@ type (
 	resourceWrapper struct {
 		resource Resource
 		timeUsed time.Time
-	}
-
-	// Setting represents a set query and reset query for system settings.
-	Setting struct {
-		query      string
-		resetQuery string
 	}
 
 	// ResourcePool allows you to use a pool of resources.
@@ -106,7 +71,6 @@ type (
 		idleTimer *timer.Timer
 		logWait   func(time.Time)
 
-		settingResources  chan resourceWrapper
 		getCount          atomic.Int64
 		getSettingCount   atomic.Int64
 		diffSettingCount  atomic.Int64
@@ -128,21 +92,6 @@ var (
 	ErrCtxTimeout = vterrors.New(vtrpcpb.Code_DEADLINE_EXCEEDED, "resource pool context already expired")
 )
 
-func NewSetting(query, resetQuery string) *Setting {
-	return &Setting{
-		query:      query,
-		resetQuery: resetQuery,
-	}
-}
-
-func (s *Setting) GetQuery() string {
-	return s.query
-}
-
-func (s *Setting) GetResetQuery() string {
-	return s.resetQuery
-}
-
 // NewResourcePool creates a new ResourcePool pool.
 // capacity is the number of possible resources in the pool:
 // there can be up to 'capacity' of these at a given time.
@@ -162,10 +111,9 @@ func NewResourcePool(factory Factory, capacity, maxCap int, idleTimeout time.Dur
 		panic(errors.New("invalid/out of range capacity"))
 	}
 	rp := &ResourcePool{
-		resources:        make(chan resourceWrapper, maxCap),
-		settingResources: make(chan resourceWrapper, maxCap),
-		factory:          factory,
-		logWait:          logWait,
+		resources: make(chan resourceWrapper, maxCap),
+		factory:   factory,
+		logWait:   logWait,
 	}
 	rp.available.Store(int64(capacity))
 	rp.capacity.Store(int64(capacity))
@@ -210,33 +158,19 @@ func (rp *ResourcePool) closeIdleResources() {
 
 	for i := 0; i < available; i++ {
 		var wrapper resourceWrapper
-		var origPool bool
 		select {
 		case wrapper = <-rp.resources:
-			origPool = true
-		case wrapper = <-rp.settingResources:
-			origPool = false
 		default:
 			// stop early if we don't get anything new from the pool
 			return
 		}
 
-		var reopened bool
 		if wrapper.resource != nil && idleTimeout > 0 && time.Until(wrapper.timeUsed.Add(idleTimeout)) < 0 {
 			wrapper.resource.Close()
 			rp.idleClosed.Add(1)
 			rp.reopenResource(&wrapper)
-			reopened = true
 		}
-		rp.returnResource(&wrapper, origPool, reopened)
-	}
-}
-
-func (rp *ResourcePool) returnResource(wrapper *resourceWrapper, origPool bool, reopened bool) {
-	if origPool || reopened {
-		rp.resources <- *wrapper
-	} else {
-		rp.settingResources <- *wrapper
+		rp.resources <- wrapper
 	}
 }
 
@@ -258,15 +192,12 @@ func (rp *ResourcePool) reopen() {
 // has not been reached, it will create a new one using the factory. Otherwise,
 // it will wait till the next resource becomes available or a timeout.
 // A timeout of 0 is an indefinite wait.
-func (rp *ResourcePool) Get(ctx context.Context, setting *Setting) (resource Resource, err error) {
+func (rp *ResourcePool) Get(ctx context.Context) (resource Resource, err error) {
 	// If ctx has already expired, avoid racing with rp's resource channel.
 	if ctx.Err() != nil {
 		return nil, ErrCtxTimeout
 	}
-	if setting == nil {
-		return rp.get(ctx)
-	}
-	return rp.getWithSettings(ctx, setting)
+	return rp.get(ctx)
 }
 
 func (rp *ResourcePool) get(ctx context.Context) (resource Resource, err error) {
@@ -281,93 +212,17 @@ func (rp *ResourcePool) get(ctx context.Context) (resource Resource, err error) 
 	// check normal resources first
 	case wrapper, ok = <-rp.resources:
 	default:
+		// now waiting
+		startTime := time.Now()
 		select {
-		// then checking setting resources
-		case wrapper, ok = <-rp.settingResources:
-		default:
-			// now waiting
-			startTime := time.Now()
-			select {
-			case wrapper, ok = <-rp.resources:
-			case wrapper, ok = <-rp.settingResources:
-			case <-ctx.Done():
-				return nil, ErrTimeout
-			}
-			rp.recordWait(startTime)
-		}
-	}
-	if !ok {
-		return nil, ErrClosed
-	}
-
-	// if the resource has setting applied, we will close it and return a new one
-	if wrapper.resource != nil && wrapper.resource.IsSettingApplied() {
-		rp.resetSettingCount.Add(1)
-		err = wrapper.resource.ResetSetting(ctx)
-		if err != nil {
-			// as reset is unsuccessful, we will close this resource
-			wrapper.resource.Close()
-			wrapper.resource = nil
-			rp.active.Add(-1)
-		}
-	}
-
-	// Unwrap
-	if wrapper.resource == nil {
-		wrapper.resource, err = rp.factory(ctx)
-		if err != nil {
-			rp.resources <- resourceWrapper{}
-			return nil, err
-		}
-		rp.active.Add(1)
-	}
-	if rp.available.Add(-1) <= 0 {
-		rp.exhausted.Add(1)
-	}
-	rp.inUse.Add(1)
-	return wrapper.resource, err
-}
-
-func (rp *ResourcePool) getWithSettings(ctx context.Context, setting *Setting) (Resource, error) {
-	rp.getSettingCount.Add(1)
-	var wrapper resourceWrapper
-	var ok bool
-	var err error
-
-	// Fetch
-	select {
-	// check setting resources first
-	case wrapper, ok = <-rp.settingResources:
-	default:
-		select {
-		// then, check normal resources
 		case wrapper, ok = <-rp.resources:
-		default:
-			// now waiting
-			startTime := time.Now()
-			select {
-			case wrapper, ok = <-rp.settingResources:
-			case wrapper, ok = <-rp.resources:
-			case <-ctx.Done():
-				return nil, ErrTimeout
-			}
-			rp.recordWait(startTime)
+		case <-ctx.Done():
+			return nil, ErrTimeout
 		}
+		rp.recordWait(startTime)
 	}
 	if !ok {
 		return nil, ErrClosed
-	}
-
-	// Checking setting hash id, if it is different, we will close the resource and return a new one later in unwrap
-	if wrapper.resource != nil && wrapper.resource.IsSettingApplied() && !wrapper.resource.IsSameSetting(setting.query) {
-		rp.diffSettingCount.Add(1)
-		err = wrapper.resource.ResetSetting(ctx)
-		if err != nil {
-			// as reset is unsuccessful, we will close this resource
-			wrapper.resource.Close()
-			wrapper.resource = nil
-			rp.active.Add(-1)
-		}
 	}
 
 	// Unwrap
@@ -379,16 +234,6 @@ func (rp *ResourcePool) getWithSettings(ctx context.Context, setting *Setting) (
 		}
 		rp.active.Add(1)
 	}
-
-	if !wrapper.resource.IsSettingApplied() {
-		if err = wrapper.resource.ApplySetting(ctx, setting); err != nil {
-			// as we are not able to apply setting, we can return this connection to non-setting channel.
-			// TODO: may check the error code to see if it is recoverable or not.
-			rp.resources <- wrapper
-			return nil, err
-		}
-	}
-
 	if rp.available.Add(-1) <= 0 {
 		rp.exhausted.Add(1)
 	}
@@ -402,14 +247,11 @@ func (rp *ResourcePool) getWithSettings(ctx context.Context, setting *Setting) (
 // This will cause a new resource to be created in its place.
 func (rp *ResourcePool) Put(resource Resource) {
 	var wrapper resourceWrapper
-	var recreated bool
-	var hasSettings bool
 	if resource != nil {
 		wrapper = resourceWrapper{
 			resource: resource,
 			timeUsed: time.Now(),
 		}
-		hasSettings = resource.IsSettingApplied()
 		if resource.Expired(rp.extendedMaxLifetime()) {
 			rp.maxLifetimeClosed.Add(1)
 			resource.Close()
@@ -419,20 +261,11 @@ func (rp *ResourcePool) Put(resource Resource) {
 	if resource == nil {
 		// Create new resource
 		rp.reopenResource(&wrapper)
-		recreated = true
 	}
-	if !hasSettings || recreated {
-		select {
-		case rp.resources <- wrapper:
-		default:
-			panic(errors.New("attempt to Put into a full ResourcePool"))
-		}
-	} else {
-		select {
-		case rp.settingResources <- wrapper:
-		default:
-			panic(errors.New("attempt to Put into a full ResourcePool"))
-		}
+	select {
+	case rp.resources <- wrapper:
+	default:
+		panic(errors.New("attempt to Put into a full ResourcePool"))
 	}
 	rp.inUse.Add(-1)
 	rp.available.Add(1)
@@ -466,7 +299,6 @@ func (rp *ResourcePool) SetCapacity(capacity int) error {
 		if oldcap == 0 && capacity > 0 {
 			// Closed this before, re-open the channel
 			rp.resources = make(chan resourceWrapper, cap(rp.resources))
-			rp.settingResources = make(chan resourceWrapper, cap(rp.settingResources))
 		}
 		if oldcap == capacity {
 			return nil
@@ -483,11 +315,7 @@ func (rp *ResourcePool) SetCapacity(capacity int) error {
 	// then we just add empty resource to the channel.
 	if capacity < oldcap {
 		for i := 0; i < oldcap-capacity; i++ {
-			var wrapper resourceWrapper
-			select {
-			case wrapper = <-rp.resources:
-			case wrapper = <-rp.settingResources:
-			}
+			wrapper := <-rp.resources
 			if wrapper.resource != nil {
 				wrapper.resource.Close()
 				rp.active.Add(-1)
@@ -502,7 +330,6 @@ func (rp *ResourcePool) SetCapacity(capacity int) error {
 	}
 	if capacity == 0 {
 		close(rp.resources)
-		close(rp.settingResources)
 	}
 	return nil
 }