@vmg 2023-09-22
This PR is now :green_apple: and ready for review. :tada:

Some notes on the development and integration of the PR:

- As you can see, the new connection pool has different enough semantics that it hasn't been feasible to implement it as a swappable replacement for the old one. I think that's perfectly OK: I tried really hard to write the LFU cache replacement to be swappable, and I don't think that any of the users ever really opted out of the new implementation because the new cache was just better in every way. Looking at all the analysis I've performed so far, I think the same thing is gonna happen here: the new pool is just strictly better for all workloads than the old one. The only reason to switch to the old one would be bugs we haven't noticed yet, which brings me to the next point...
- Obviously this shouldn't go into Vitess 18 which is being frozen next week. We'll merge after the release branch is cut and we have all the Vitess 19 period to iron out all the bugs and kinks in the new implementation. We'll be able to ship this pool in Vitess 19 with high confidence.
- I've found several bugs in tests regarding broken connections which are not related to this PR, but were surfaced by it, because we now re-use connections in a LIFO manner. I've extracted those fixes into a separate PR which has already been merged: https://github.com/vitessio/vitess/pull/14076