@vmg 2023-09-22
One last significant bug that is handled by this PR, and which was present in the old connection pool implementation:

Because of the way the MySQL protocol works (it is client-led), there's no way to detect when the MySQL server has closed a connection from its side. So there's this issue where we can pull a connection from the pool that has been sitting there idle for a while, and the connection is broken because there was a network flake or the MySQL server just killed us. However, we won't know that the connection is broken until we actually try to perform a query on it, and we get an actual error.

Most of the time, this is well handled in Vitess because our default implementation for a pooled connection has transparent retries: when you attempt a query and find out that he connection is broken, we'll automatically attempt a reconnection and retry the query.

But there are times when this is not an option: when using a `StatefulConnection` in the `vttablet`, we cannot perform any reconnection/retries, because the connection represents a consistent state (often an open transaction) and re-connecting it would break that invariant and lose our transaction. Whenever a query fails in such connection, we cannot afford to retry-it (even if it's the very first statement in the transaction, because the error could have happened _after_ the statement was executed).

This leaves us in a situation when sometimes a `StatefulConnection` just fails right away because it pulled a broken connection from the pool, and we have to return an error to the user. This situation has always been present, but it's become more apparent now in CI because the new LIFO pool finds more of these broken connections, particularly in the healthcheck endtoend tests, which often include killing or restarting `mysqld` instances to see how the tablet reacts.

So how do we fix this? Fortunately I've been through this before: I found the same issue when deploying GitHub's first production service in Go a few years ago. The OSS MySQL driver for Go (`go-mysql-driver`) suffered from the same problem of returning broken connections from the pool. I devised an ingenious approach to figure out whether the connection is broken by skipping the Go socket abstraction and performing directly a non-blocking read on the socket.

You can read a deep dive into the problem and the fix in this blog post I wrote: https://github.blog/2020-05-20-three-bugs-in-the-go-mysql-driver/

I've implemented the same fix in Vitess' `mysql.Conn` and lo & behold, our `StatefulConnection` can no longer fail when they pull a broken connection from the pool. This removes a _good chunk_ of flaky tests from CI and I'm sure will improve reliability when using the new connection pool in busy `mysqld` instances.