@harshit-gangal 2023-10-05
The changes are looking good.

But for some reason, benchmarks are not as expected.
comparing this PR commit with 18-rc1 tag commit showing some regression
https://benchmark.vitess.io/compare?ltag=6ab165ade925b35a00cf447827d874eba13998b6&rtag=ae288ec9175e456c811784cf1a2944eb436a257b&ptagM=0

I will work with @frouioui  to ensure all the benchmark results are accurate.